+++ src/main/java/agents/forecast/PriceForecasterApi.java
@@ -85,8 +85,8 @@ public class PriceForecasterApi extends MarketForecaster {
 
 		call(this::logClearingPrices).onAndUse(DayAheadMarket.Products.Awards);
 		call(this::registerClearingTime).onAndUse(DayAheadMarket.Products.GateClosureInfo);
-		call(this::sendPriceForecast).on(Forecaster.Products.PriceForecast)
-				.use(ForecastClient.Products.PriceForecastRequest);
+		call(this::sendPriceForecast).on(DamForecastProvider.Products.PriceForecast)
+				.use(DamForecastClient.Products.PriceForecastRequest);
 	}
 
 	/** Extracts and store power prices reported from {@link DayAheadMarket}