+++ src/main/java/agents/trader/electrolysis/ElectrolysisTrader.java
@@ -10,8 +10,8 @@ import agents.electrolysis.Electrolyzer;
 import agents.electrolysis.ElectrolyzerStrategist;
 import agents.flexibility.BidSchedule;
 import agents.flexibility.Strategist;
-import agents.forecast.ForecastClient;
-import agents.forecast.Forecaster;
+import agents.forecast.DamForecastClient;
+import agents.forecast.DamForecastProvider;
 import agents.forecast.MarketForecaster;
 import agents.markets.DayAheadMarket;
 import agents.markets.DayAheadMarketTrader;
@@ -88,11 +88,11 @@ public class ElectrolysisTrader extends FlexibilityTrader implements FuelsTrader
 		fuelData = new FuelData(fuelType);
 
 		call(this::prepareForecasts).on(Trader.Products.BidsForecast).use(MarketForecaster.Products.ForecastRequest);
-		call(this::requestElectricityForecast).on(ForecastClient.Products.PriceForecastRequest)
+		call(this::requestElectricityForecast).on(DamForecastClient.Products.PriceForecastRequest)
 				.use(DayAheadMarket.Products.GateClosureInfo);
 		call(this::requestHydrogenPriceForecast).on(FuelsTrader.Products.FuelPriceForecastRequest)
 				.use(DayAheadMarket.Products.GateClosureInfo);
-		call(this::updateElectricityPriceForecast).onAndUse(Forecaster.Products.PriceForecast);
+		call(this::updateElectricityPriceForecast).onAndUse(DamForecastProvider.Products.PriceForecast);
 		call(this::updateHydrogenPriceForecast).onAndUse(FuelsMarket.Products.FuelPriceForecast);
 		call(this::prepareBids).on(DayAheadMarketTrader.Products.Bids).use(DayAheadMarket.Products.GateClosureInfo);
 		call(this::digestAwards).onAndUse(DayAheadMarket.Products.Awards);